requirements.txt
langchain>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.2.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
